+++ src/tests/unit/mocks/mockAdapter.ts
@@ -65,6 +65,7 @@ const implementedMethods: ImplementedMethodDictionary<ioBroker.Adapter> = {
 	sendTo: "no error",
 	sendToHost: "no error",
 	getHistory: "normal",
+	// @ts-expect-error This method was deprecated
 	setBinaryState: "normal",
 	getBinaryState: "normal",
 	getEnum: "normal",
@@ -217,9 +218,9 @@ export function createAdapterMock(
 			},
 			...args: any[]
 		) => {
-			const callback = getCallback<ioBroker.GetObjectListCallback>(
-				...args,
-			);
+			const callback = getCallback<
+				ioBroker.GetObjectListCallback<ioBroker.Object>
+			>(...args);
 
 			if (typeof callback === "function") {
 				let objects = values(db.getObjects("*"));